fixed sorting url
