+++ src/components/PeopleRender.tsx
@@ -6,7 +6,6 @@ import { EmptyTableMessage } from './EmptyTableMessage';
 import { Person } from '../types';
 import { Link } from 'react-router-dom';
 import { useFilters } from '../hooks/useFilters';
-import { getSearchWith } from '../utils/searchHelper';
 
 interface Props {
   people: Person[];
@@ -15,26 +14,7 @@ interface Props {
 const SORT_ORDERS = ['Name', 'Sex', 'Born', 'Died'];
 
 export const PeopleRender: React.FC<Props> = ({ people }) => {
-  const { searchParams, order, sort } = useFilters();
-
-  const handleSorting = (sortBy: string | null) => {
-    const isSortSameAsArgument = (sort === sortBy) !== null;
-    const lowerSortBy = sortBy ? sortBy.toLowerCase() : '';
-
-    if (!sort) {
-      return getSearchWith(searchParams, { sort: lowerSortBy });
-    }
-
-    if (sort && !order && isSortSameAsArgument) {
-      return getSearchWith(searchParams, { sort: lowerSortBy, order: 'desc' });
-    }
-
-    if (sort && order && isSortSameAsArgument) {
-      return getSearchWith(searchParams, { sort: null, order: null });
-    }
-
-    return '';
-  };
+  const { handleSorting } = useFilters();
 
   return (
     <>
