+++ src/hooks/useFilters.ts
@@ -1,4 +1,5 @@
 import { useSearchParams } from 'react-router-dom';
+import { getSearchWith } from '../utils/searchHelper';
 
 export const useFilters = () => {
   const [searchParams, setSearchParams] = useSearchParams();
@@ -10,6 +11,25 @@ export const useFilters = () => {
   const order = searchParams.get('order');
   const hasAnySearchParam = sexInSearch !== null || !!query || cents.length > 0;
 
+  const handleSorting = (sortBy: string | null) => {
+    const lowerSortBy = sortBy ? sortBy.toLowerCase() : '';
+    const isSortSameAsArgument = sort === lowerSortBy;
+
+    if (!sort || sort !== lowerSortBy) {
+      return getSearchWith(searchParams, { sort: lowerSortBy, order: null });
+    }
+
+    if (sort && !order && isSortSameAsArgument) {
+      return getSearchWith(searchParams, { sort: lowerSortBy, order: 'desc' });
+    }
+
+    if (sort && order && isSortSameAsArgument) {
+      return getSearchWith(searchParams, { sort: null, order: null });
+    }
+
+    return '';
+  };
+
   return {
     searchParams,
     setSearchParams,
@@ -19,5 +39,6 @@ export const useFilters = () => {
     sort,
     order,
     hasAnySearchParam,
+    handleSorting,
   };
 };
